Change the mechanism that pulld and push use to announce changes.
The mechanism used to be a request to /pullpullpull on the pulld server, but since we are
locking down external IP address this won't work anymore when using 'pushcli'. Switch to
using a change in instance level metadata to trigger pulld to go look for new packages to
install.

See:
  https://cloud.google.com/compute/docs/storing-retrieving-metadata#waitforchange

BUG=skia:

Change-Id: I62f4fa6e4ce926a2e2a5ccab8332d10264bdc862
Reviewed-on: https://skia-review.googlesource.com/7306
Commit-Queue: Joe Gregorio <jcgregorio@google.com>
Reviewed-by: Ravi Mistry <rmistry@google.com>