Limit returned list by failures handler
Currently there is a huge number of 'failed' digests
in the failure database, cause the failure list handler to be very slow.
This trims the number of failures reported down to the last 50 failures. 
Once the fake failures are removed the number should be much smaller 
since we have stricter filtering for digests in place. 

BUG=skia:

Change-Id: I52b3daf061cf4f57fe93ddd3c9fbb4ba6044b02e
Reviewed-on: https://skia-review.googlesource.com/8108
Reviewed-by: Stephan Altmueller <stephana@google.com>
Commit-Queue: Stephan Altmueller <stephana@google.com>