Fiddle: Clean up the embedded iframe view.
This will make it possible to embed fiddles directly
in docs so that code and output are always in sync.
Also, all the fiddles are runnable inline.

Here is an example before:

  https://screenshot.googleplex.com/GWXEb2ykcr6.png

And after:

  https://screenshot.googleplex.com/OdFjNtyENX7.png

This also allows embedding text-only fiddles.

A future CL will address the sizing issue, i.e.
right now the iframes may not be sized correctly
by default.

BUG=skia:

Change-Id: Ide5f26bf48726c22dce893f3920a8c3cbf856b06