Backend of search re-write in Gold
This is the first CL that rewrites search in Gold.
The UI part will follow shortly. 

This implements functionality to search across the current tile,
to be used in parallel to current search function. 
It is exposed at a new HTTP endpoint, so we can 
use it in parallel to the existing search until 
it's feature complete. At that point the current search 
will be removed. 

It establishes a pipeline that allows to filter not 
only across traces (bot runs), but also across results
of the diff operation. The primary usecase is to triage 
many slightly diverging images fast. 

It adds a new system for reference points (closest
positive, negative etc.) that should be easy to extend.

BUG=skia:

Change-Id: Ief792cb66837f84f1fe3a1d2248868937bc1c0ec
Reviewed-on: https://skia-review.googlesource.com/8386
Commit-Queue: Stephan Altmueller <stephana@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>