Update patch set 1
Patch Set 1: Code-Review+1

I think we should figure out how relative priorities will compare in terms of scheduling, eg. a task spec with priority p1 and k1 commits in the blamelist will schedule at the same priority as a task spec with priority p2 and k2 commits in the blamelist.  Otherwise, we'd just be choosing priorities arbitrarily.

For untested commits, the score is p*t*(n + 1 - 1/n), where n is the number of commits in the blamelist, t is the time decay factor for the commit, and p is the priority, so if we assume the time decay factor is equal we can do:
p1*(n1 + 1 - 1/n1) = p2*(n2 + 1 - 1/n2)

Patch-set: 1
Reviewer: Eric Boren <5015@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Code-Review=+1