Rewrite search front-end
This adds a refactored version of the current search frontend to 
temporarily give us accessto both searches until the new version is 
complete enough to replace the old one. 

Introduces 
    newdigest-details-sk 
    newsearch-page-sk 
elements which are modified copies of digest-details-sk and search-page-sk.

We extend the filtering dialog to expose the previously 
added backend features. 

BUG=skia:

Change-Id: I77e2adb00930e2f95060e5317b94367c53590623