Add flakiness tool
This is mainly a proof of concept, focusing on finding flakes. The idea
is to keep it extensible so that we can easily add new patterns which
might indicate flakiness.

In the future this should be extended with a UI, create unique
identifiers for flakes based on component tasks, then add support for
triaging flakes, associating them with bugs, marking as ignore, etc.

BUG=skia:6006

Change-Id: I2ac5100b093c9c7fe2a8ed79230d7668c8b64268