Update patch set 1
Patch Set 1:

(1 comment)

In order for this to work, you'll need to change ComputeBlamelist to not return empty for a non-HEAD revision.

I'm also not certain how this would work if I cherry-pick my new task to a different branch.

Alternative suggestion: In ComputeBlamelist, check if the task spec exists at each repo state visited in Recurse. If not, stop recursing. This will cause the blamelist to be calculated correctly, allowing the new task to run at HEAD and allowing bisects if there are enough bots free. Presumably the task specs are cached and this wouldn't cause a huge slowdown; however, we would need to be smarter about weekly tasks -- either reduce MAX_BLAMELIST_COMMITS or stop recursing if we encounter a commit before our scheduling window.

Patch-set: 1
Reviewer: Ben Wagner <5120@5086850b-e599-37d2-8ec8-ff16aeccbf02>