Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> In order for this to work, you'll need to change ComputeBlamelist to not return empty for a non-HEAD revision.
> 

With this solution, we don't call ComputeBlamelist any more for new task specs.

> I'm also not certain how this would work if I cherry-pick my new task to a different branch.
> 

I think we'd hit the max blamelist size and default to a single-commit blamelist.  FTR it'd be the same behavior before this change.

> Alternative suggestion: In ComputeBlamelist, check if the task spec exists at each repo state visited in Recurse. If not, stop recursing. This will cause the blamelist to be calculated correctly, allowing the new task to run at HEAD and allowing bisects if there are enough bots free. Presumably the task specs are cached and this wouldn't cause a huge slowdown; however, we would need to be smarter about weekly tasks -- either reduce MAX_BLAMELIST_COMMITS or stop recursing if we encounter a commit before our scheduling window.

That'd definitely be more correct.  We'd have to pass in a TaskCfgCache to ComputeBlamelist (which might mess up usage in datahopper_internal?) and I'm concerned about how slow the locking in the cache would be.

Patch-set: 1
Reviewer: Eric Boren <5015@5086850b-e599-37d2-8ec8-ff16aeccbf02>