Update patch set 2
Patch Set 2:

I disagree with this revert. Spam will be less noticeable once we thread in a single conversation. If the main reason for revert is reboot alerts than that does not happen that often anyway. Plus it is more important to have silences apply within a few seconds.

Patch-set: 2