Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)

Hrm.  I was just playing with this, with the aim of adding the initial commit as the highest priority but allowing subsequent candidates to follow the normal flow, but I realized that we may not even have the commit which introduced the task spec in the set of candidates, since a required dependency may not be met at that commit.

I kind of want a TaskSpec.IntroducedIn property which we could use in ComputeBlamelist.  If it's a new task spec, instead of cutting off at one commit in the blamelist, we could trace back until we get to the IntroducedIn commit.

Patch-set: 1
Reviewer: Eric Boren <5015@5086850b-e599-37d2-8ec8-ff16aeccbf02>