Update patch set 2
Patch Set 2:

I may not be understanding the full context, but it seems that you're planning to switch from timeouts being an error to post-processing the test times and recategorizing the tests based on previous runs.

Would it be simpler to just parse the test times that 'go test' outputs? It claims to use a standard format:

"...It prints a summary of the test results in the format:

	ok   archive/tar   0.011s
	FAIL archive/zip   0.022s
	ok   compress/gzip 0.033s
	...
"

Patch-set: 2
Reviewer: Ben Wagner <5120@5086850b-e599-37d2-8ec8-ff16aeccbf02>