[status] Add find_breaks package
This is intended to be the first stage in a pipeline
which finds groups of related failures and distills
them into actionable alerts for sheriffs.

This stage finds failed tasks and groups them into
related failures, following similar logic to what a
human might do while looking at the status page to
determine which breakages are cased by which commits.

BUG=skia:

Change-Id: I1b52add197eba547010d65b0d6d4b2ad078e40b3