Update patch set 2
Patch Set 2: Code-Review+1

(2 comments)

I'm not clear on what "child" and "parent" mean in this context.

Based on your description and the code in this CL, "child" has a DEPS file that points to "parent." But that seems backwards from how I'm used to thinking about it, e.g. I would consider Skia to be a child of Chromium, not the other way around.

I'm guessing we may also want to go the other way, similar to how we have the ability to run Chromium tryjobs on Skia CLs. I don't think this CL allows that -- see comment below.

Patch-set: 2
Reviewer: Ben Wagner <5120@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Code-Review=+1