Update patch set 4
Patch Set 4:

(2 comments)

> Patch Set 2: Code-Review+1
> 
> (2 comments)
> 
> I'm not clear on what "child" and "parent" mean in this context.
> 
> Based on your description and the code in this CL, "child" has a DEPS file that points to "parent." But that seems backwards from how I'm used to thinking about it, e.g. I would consider Skia to be a child of Chromium, not the other way around.
> 
> I'm guessing we may also want to go the other way, similar to how we have the ability to run Chromium tryjobs on Skia CLs. I don't think this CL allows that -- see comment below.

The "parent" repo has a DEPS file which includes the "child" repo.  This CL is intended to enable running trybots defined in the internal repo on a Skia CL, so internal_test.git is the parent repo (having Skia in its DEPS file) and skia.git is the child.

Triggering the try job as specified above sends the job to the task scheduler instance for the parent repo (internal_test.git in this example).  FTR this is exactly how Chromium trybots are triggered for Skia CLs.  We just need to plumb the try_job_repo through so that bot_update puts the patch in the right place, and pass the correct patch_repo along to the recipe, which will also need to be responsible for syncing the correct parent/child+patch repo state.

Patch-set: 4
Reviewer: Eric Boren <5015@5086850b-e599-37d2-8ec8-ff16aeccbf02>