[fuzzer] Fix rolling logic
I missed a plumbing step on the front end side.
Made backend more resistent to 404 errors.
It appears that by using httputils.NewBackOffTransport(), storage.ErrObjNotExist
is not thrown because the API doesn't know how to deal with its errors.
To compensate, we check for 404 in the error message because BackoffTransport
includes it in the message.

Also tweak the release to work, since the css was removed.

Bug: skia:6094
Change-Id: I753cd4422fb30d44a5f8ebfbfbd592a6ab8ffe9c