Fix hotspare disk space leak
Basically, rsync was overwriting a file that was in use
and this caused the old bytes to be leaked.

This means the spare won't be serving for the duration of
the rsync command, which is about 15 seconds.

The RPIs are pretty robust and I've seen them go up to a
minute w/o any master and still be fine.

I think this behavior is left over from when the network was
really really slow and rsync could take a few minutes.

Bug: skia:
Change-Id: I50b21c4d810f6da0e381100f416ed4fe2c6fe1e4
Reviewed-on: https://skia-review.googlesource.com/14440
Reviewed-by: Stephan Altmueller <stephana@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>