Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1:
> > > > 
> > > > (1 comment)
> > > 
> > > I see, so you're actually in a case like:
> > > 
> > >  ...- a - b - c - d - e (branch)
> > >         \       \   \
> > >  ...- g - h - i - j - k (master)
> > > 
> > > where all commits except e are ancestors of k, but we want to ignore anything that's an ancestor of e, including those that are ancestors of k.
> > > 
> > > What you have is fine.
> > 
> > So the problem is that, if master is blacklisted, we'll never test anything but e, even though we do want to test all commits on "branch".
> 
> Sorry, I should have labeled e as "skia-master" and k as "development branch in internal repo".
> 
> But PS1 is good to commit.

I'm actually not convinced that this is correct enough to land.  Sure, it works for the current layout, but if we fork a new branch from master in skia.git and immediately blacklist it, we're going to stop testing everything in master before and up to the branch point.

Patch-set: 1
Reviewer: Eric Boren <5015@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Commit-Queue=0