Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Try jobs failed on following builders:
>   Infra-PerCommit-Large on skia.primary (JOB_FAILED, build has not started yet; builder either lacks capacity or does not exist (misspelled?))
>   Infra-PerCommit-Medium on skia.primary (JOB_FAILED, build has not started yet; builder either lacks capacity or does not exist (misspelled?))
>   Infra-PerCommit-Small on skia.primary (JOB_FAILED, build has not started yet; builder either lacks capacity or does not exist (misspelled?))
> 
> Bot data: {"action": "cancel", "triggered_at": "2017-05-10T14:41:28.0Z", "revision": "730378610c83acea9bb7815519b713e6ce35ecba"}

The patch failed to apply on the task scheduler.  Normally we get an explicit cancellation by the scheduler, but this is a different error.  I'm not sure if this is just the new error message for the canceled state, or if something else has gone wrong here.

Patch-set: 3
Reviewer: Eric Boren <5015@5086850b-e599-37d2-8ec8-ff16aeccbf02>