Make indexing across main traces and ignored consistent
For each tile (test results of the last 100 commits) 
we build an index to search faster. This fixes the 
issue where some parts of the index were created with 
the ignores and some where created without the ignores.

We now calculate it for both and make it a flag to most 
functions in SearchIndex whether it should include 
the ignored. 

As a side effect, the warmer should also more 
aggressively prefetch and diff ignored digests/images.

Bug: skia:
Change-Id: I069a77820860cf4f32ae683ba6d2ba6013fd5d04