Make promalertsclient use alerts/groups
dispatch.APIAlert ends up being a not so good type because it embeds
a vendored type. This means we can't use it (at least, not if we want to
make one ourselves).

We base our alert of of model.Alert, which doesn't have any info on silenced
so we include it.

Additionally the /alerts api doesn't (yet) tell us if an alert is silenced, so
we need to use /alerts/groups instead.

Bug: skia:
Change-Id: I32a08f22cc4d7c4debe301fc4f0fc3dc7e59af49
Reviewed-on: https://skia-review.googlesource.com/18060
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Ravi Mistry <rmistry@google.com>