[gold] Enable diff server
Changes configs so that all diffing is handled
by the diff server and images are served directly
from the diff server.

Increases the grpc message size of the GRPC service
between Gold and the diffserver which was found o the be 
an issue during testing.

Bug: skia:
Change-Id: Ib6a40a5cd657751ead09182bb2f0efb51cc624ad
Reviewed-on: https://skia-review.googlesource.com/19980
Commit-Queue: Stephan Altmueller <stephana@google.com>
Reviewed-by: Ravi Mistry <rmistry@google.com>