Update patch set 2
Patch Set 2:

After making the files, server can be run by command like: 
$ skiasilver -service_account_file=/usr/local/google/home/lchoi/service-account.json -no_cloud_log=true -alsologtostderr=true -local=true -diff_server_grpc=:9000 -diff_server_http=http://127.0.0.1:9001

I did set up a route using the ImageHandler function, but I'm unsure how to leverage it to serve images like Gold does. You'll see in my diff-details-sk page that as of now I'm just returning explicit image paths to the local diff server that I have running in the background. 

The biggest question I have now after reading @stephana's explanation of image serving on dynamite is whether I should be using an instance of mem_diffstore or net_diffstore for local testing, and how this will affect the paths for images when the server is actually landed in production. To clarify, my current understanding is that based on the recent changes to skia_org_nginx, I'll be able to access the production diff server directly through the /img/ prefix, but I am still confused about how I could replicate this type of path when developing and testing locally.

Moreover, how images and diffs get to the point that they are able to be served by the diff server is quite a bit unclear to me. Is it simply the case that once uploaded to GS, the images are available to the diff server, or are there intermediate steps? And as for the diff images, where is the actual call that takes in digests, creates the visual result of the diff, and stores this result in img/diffs taking place? I assume it's somewhere within the logic of the diff server, as @stephana mentioned all the image diffing has been factored out of Gold, but I can't seem to locate it. 

Sorry for all the questions, and thanks so much for your continued help!

Patch-set: 2
Reviewer: Stephan Altmueller <5620@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Reviewer: Ravi Mistry <5145@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Reviewer: Lawrence Choi <7450@5086850b-e599-37d2-8ec8-ff16aeccbf02>