Update patch set 5
Patch Set 5:

Stephan,

I wasn't sure how to go about changing the gRPC implementation of the DiffService to match the new DiffStore interface. Specifically, I don't understand how the results of the DiffFn would be accessible in diffservice_impl, as the DiffStore that the DiffService contains is not guaranteed to be a memDiffStore. I did add type assertions all over the existing codebase to account for the new interface, including inside diffservice_impl, and all the tests in diffstore_test do pass. Given that TestNetDiffStore in this file calls NewDiffServiceServer and creates a netDiffStore instance using this server, is this indicative of the gRPC implementation working correctly? If not, could you elaborate on the comments you made above regarding serializing DiffFn results, and explain how to test the correctness of the gRPC implementation? Thank you!

Patch-set: 5
Reviewer: Lawrence Choi <7450@5086850b-e599-37d2-8ec8-ff16aeccbf02>