Customizable image path configuration for DiffStore
MemDiffStore and ImageLoader objects can now be instantiated with the following
arguments:
-gsImageBaseDirs: string that represents the GS image path after the GS bucket,
but not including the directory that specifies the run
-gsImageRunDir: string that represents the GS directory specifying the run id,
usually in the format "{userid}-{timestamp}". May be empty if images are not
stored in such a directory, as in the case of Gold.
-usesDigests: bool that specifies whether the images in GS are stored as hash
digests. Used to bypass digest validation if not needed and extend the existing
image storage logic with a hyphenated structure, details below.

Flags were added to main.go of both skiacorrectness and skia_diff_server to
allow for client-side configuration of these fields.

If usesDigests is true, MemDiffStore and ImageLoader will assume all the images
are stored in one GS directory. If not, these objects will parse the URL request
made to the diff server, split the requested image name on hyphens, and recreate
the directory structure in GS as if the hyphens were '/' characters. To give
concrete examples:

$ skia_diff_server -alsologtostderr=true -grpc_port=:9000
-image_dir="/usr/local/google/home/lchoi/data/diff_server/imagediffs"
-image_port=:9001 -no_cloud_log=true -prom_port=":23000"
-service_account_file=/usr/local/google/home/lchoi/service-account.json
-log_dir="/usr/local/google/home/lchoi/data/diff_server/logs"

will create a diff server that has the default values of the abovementioned
flags, which are "dm-images-v1", "", and true, respectively. Then, a request
to the diff server like
"http://127.0.0.1:9001/img/images/000058effb1677e6cefca0f7ec3ed3d0.png"
will successfully load the image just like Gold currently does.

Running
$ skia_diff_server -alsologtostderr=true -grpc_port=:9000
-image_dir="/usr/local/google/home/lchoi/data/diff_server/imagediffs"
-image_port=:9001 -no_cloud_log=true -prom_port=":23000"
-service_account_file=/usr/local/google/home/lchoi/service-account.json
-log_dir="/usr/local/google/home/lchoi/data/diff_server/logs"
-gs_buckets="cluster-telemetry" -gs_subdirs="tasks/benchmark_runs"
-gs_rundir="rmistry-20170623184523" -uses_digests=false

will allow you to make requests to the diff server like
"http://127.0.0.1:9001/img/images/nopatch-http___google_co_ao.png", with the
nopatch prefix representing the directory structure in GS following
gsBucket/gsImageBaseDirs/gsImageRunDir. Multiple hyphens can be added to the
request to match the image location in GS, if needed.

I thought that using hyphens in the image request was a sound approach as these
same hyphenated filenames are used to create the filenames of the locally stored
images. Without the prefixes to specify inner directories, requests to the
nopatch/ and withpatch/ folders would overwrite each other with the existing
implementation, as screenshots of a web page with and without a patch
are stored with the same name in GS.

Change-Id: I0f36da7247ceb4bc4d1b9621b589c8dade86313f