Abstract out diff function from diffstore
diffFn signature created to remove hardcoded diff function and
facilitate plugging in of custom diff functions for other use cases

memDiffStore can now be instantiated with a custom diffFn that takes
in two images and returns an application dependent diff result along
with a diff image. If the diffFn is not specified, it defaults to
metrics.DefaultDiffFn.

Change-Id: I2ecbd92a7a5a938b8fc81361d48064eba74f715f
Bug: skia: