Update patch set 11
Patch Set 11:

> Patch Set 10: Code-Review+1 Commit-Queue+2
> 
> > Patch Set 10:
> > 
> > This seems to make tests pass on MariaDB. However, a better approach might be to run tests on CloudSQL.
> 
> IMO this is a good fix regardless since the behavior is undefined.  I'm going to land this so that I can move on with recreating the GCE instances.

Joe and/or Stephan, please also take a look. My guess is that this change is OK, but there is the potential that it could cause much more data to be read from the DB than before. Presumably there won't be very many rows in exp_test_change that match on name and digest, though.

Patch-set: 11