Update patch set 5
Patch Set 5:

Latest patch is dependent on the latest upload to the image paths CL: https://skia-review.googlesource.com/c/20770/

Added a jobs page and a jsonJobsHandler function that returns a list of CT Pixel Diff runs, as found in GS. For now, the jsonJobsHandler just returns a hardcoded list consisting of 2 of Ravi's test runs. A user can click on a specific job to see the screenshots and diff images corresponding to that run. 

I have two related questions in mind as we move forward. Firstly, I'm going to need to be able to access something like a full list of all CT Pixel Diff runs in order to return the abovementioned list in jsonJobsHandler to the user. I assume the first step is making a separate directory for Pixel Diff runs and storing each run as a separate subdirectory there, but how do we go about aggregating all these past runs so that this information is accessible to the new server? Is there going to be some top-level metadata that lists the {userid-timestamp} for each of the runs, or should we rely on some other approach? Secondly, I'm going to need to be able to access the metadata for a specific run in jsonDiffHandler in order for the server to know the filenames of all the screenshots taken (explained in Patch Set 4). What is the best approach for obtaining this metadata from GS? Should I implement something like a downloadMetadata function in mem_diffstore?

Patch-set: 5
Reviewer: Lawrence Choi <7450@5086850b-e599-37d2-8ec8-ff16aeccbf02>