Abstract out diff function from diffstore
diff.DiffFn signature created to remove hardcoded diff function and facilitate
plugging in of custom diff functions for other use cases

MemDiffStore can now be instantiated with a custom DiffFn that takes in
two images and returns an application dependent generic interface along with a
diff image. If not specified, the diffFn will default to metrics.DefaultDiffFn.

gRPC implementation of DiffService refactored to accomodate the genericized
DiffStore.Get(). GetDiffsResponse is now represented with a byte array, and
NewDiffServiceImpl and NewNetDiffStore now require an additional parameter, an
implementation of the util.LRUCodec interface that facilitates conversion
between byte slices and maps of the custom metrics struct, of type
map[string]interface{}. This implementation, MetricMapCodec, was added to
diffservice_impl.go and includes Encode() and Decode() functions that enable
the serialization/deserialization of the results of the DiffFn.

Change-Id: I2ecbd92a7a5a938b8fc81361d48064eba74f715f
Bug: skia: