silveringestion and dmstore packages
silveringestion enables the creation of a silverProcessor, which implements
the ingestion.Processor interface and facilitates the continuous
parsing of metadata from GS, from which information about screenshots taken
during CT pixel diff runs can be extracted. A silverProcessor contains a boltDB
instance that holds information about screenshots taken: each run has its own
bucket, and within a bucket, there are records that have screenshot filenames
as keys and a list of two bools as values. The bools correspond to whether
the nopatch and withpatch images for a particular screenshot filename have been
processed. Once both are marked true, the processor uses its DMStore to calculate
diff metrics for the two images.

DMStore contains a DiffStore object to calculate diff metrics and a boltDB
instance to hold them. The Add() function takes in a run ID and a screenshot
filename to generate the GS paths for the nopatch and withpatch images. The
DiffStore is then used to calculate the diff metrics for these images, which
is stored in the boltDB instance. In the instance, each run has its own bucket,
and the records have screenshot filenames as keys and a serialized
diff.DiffMetrics object as values.

Bug: skia:
Change-Id: Ice71972f8117dfc6c25cd4f52a9185710df3b66c