Customizable image path configuration for DiffStore (new version)
ImageLoader and MemDiffstore refactored so that images (items under img/images)
can be accessed  through explicit paths that represent the GS directory structure
following the gsBucket and gsBaseDir that were supplied to instantiate the
ImageLoader. The functions ImageHandler, IsOnDisk, PurgeImages, imageLoadWorker,
and saveImgInfoAsync now check if the supplied string (the substring that occurs
after img/images/ in the URL request) contains a "/" character to see whether
the string represents a filename or a filepath. In the former case, images are
saved, accessed, removed, etc. through the use of fileutils.TwoLevelRadixPath,
as in the case of Gold. In the latter case, images are handled by joining the
ImageLoader's localImgDir with the given string representing a path; createPath
and saveFilePath were added to common.go to facilitate this functionality. In
order to give a few concrete examples, I also modified main.go of skia_diff_server
to include a gsBaseDirs flag that is used to create the instance of the MemDiffStore.

Running
$ skia_diff_server -alsologtostderr=true -grpc_port=:9000 -image_port=:9001
-no_cloud_log=true -prom_port=":23000"
-service_account_file=/usr/local/google/home/lchoi/service-account.json
-gs_buckets="skia-infra-gm,chromium-skia-gm" -gs_basedirs="dm-images-v1"

will successfully load image requests to the diff server like
"http://127.0.0.1:9001/img/images/000058effb1677e6cefca0f7ec3ed3d0.png", and the
image will be downloaded to tmp/imagedir/images/00/00/000058effb1677e6cefca0f7ec3ed3d0.png,
just like Gold currently does.

On the other hand, running
$ skia_diff_server -alsologtostderr=true -grpc_port=:9000 -image_port=:9001
-no_cloud_log=true -prom_port=":23000"
-service_account_file=/usr/local/google/home/lchoi/service-account.json
-gs_buckets="cluster-telemetry" -gs_basedirs="tasks/benchmark_runs"

will set up a server that is able to access Ravi's test CT pixel diff runs, located
in  cluster-telemetry/tasks/benchmark_runs in GS. Then, requests to the diff server
like "http://127.0.0.1:9001/img/images/rmistry-20170623184523/withpatch/http___www_google_com.png"
will successfully download and load the image. The image is saved to
tmp/imagedir/images/rmistry-20170623184523/withpatch/http___www_google_com.png.

MemDiffStore further modified to allow for customized configuration of how
diff images and diff metrics are stored. A MemDiffStore instance now has three
new fields: one is a function that creates the BoltDB diffId key for two images,
the second is a function that creates the diff image filename for two images, and
the third is a function that takes in a diffId to retrieve the two image filenames
that were used to create that diff. If not specified, these functions will
default to combineDigests, getDiffImgFileName, and splitDigests, respectively.
The first two implement the GetDiffString function signature, while the third
implements the GetLeftAndRight function signature. The new pixel diff server will
utilize new implementations of these signatures called GetCommonRunUrl,
GetCommonRunUrlImgName, and GetNoAndWithPatch to insantiate its DiffStore; this
will enable divergence from the way Gold currently stores diff results, which
yields unneccesarily long diff keys and filenames on the new server. DiffIds for
BoltDB on the new server will now be specified by strings of the form "runID:URL",
diff images will be saved under tmp/imagedir/diffs/runID/URL.png, and the two
image paths can be retrieved from a diffID by splitting on the colon and either
adding a "nopatch" or "withpatch" in between the runID and the URL.

Bug: skia:
Change-Id: I80aa73aebdc5bac0778f7e1ce6eaa753e39920b2