Update patch set 4
Patch Set 4:

I was able to get rid of most of the occurrences of String.Index(s, "/") by using the new DiffPath and ImagePath functions. However, I didn't know how to leverage the new interface in ImageHandler, which should only validate the digest in the path in the URL request and rewrite the path to include the radix prefix if the client is Gold. For now, ImageHandler is still checking for "/" characters in the request in order to determine if it should do the validation and path rewrites. 

Also, I was able to delete most of the Gold custom functions like combineDigests, getDiffBasename, saveFileRadixPath, etc. by folding them into the GoldIDPathMapper implementation but metricsRec.IndexValues explicitly uses splitDigests. I tried passing in the mapper through newMetricsStore and having a global variable to hold it in metrics_store so that IndexValues could use the specified mapper.SplitDiffID function, but testing seemed to indicate that there are times when IndexValues is called before newMetricsStore, causing this approach to fail with a null pointer exception. Don't think this will affect the new server, as it will have its own boltDB instance from which to query diff metrics, but would be nice if there was some way to completely flesh out the abstraction and remove all the custom Gold functions.

Patch-set: 4
Reviewer: Lawrence Choi <7450@5086850b-e599-37d2-8ec8-ff16aeccbf02>