Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (8 comments)
> 
> I wrote a few inline comments and you should still consider them, but first please think about my suggestion about a different architecture. 
> 
> The packages should be structured differently IMO.
> At the bottom of your stack should be the 'dmstore' which does nothing but storing diff results and make them query-able.
> Each record in this store should be whatever your diff metric struct is.
> 
> On top of that sits the Ingester. It continuously ingests files, parses them and calculates diff metrics using a DiffStore.
> It then stores those diff metrics in its DMStore instance. 
> 
> Also on top of the same DMStore instance are the HTTP handlers that expose the query api to the front-end. They query the DMStore instance to extract results based on CT runs.
> 
> I would suggest to detach this from the main function for now and make the master git branch the upstream and leave out the changes to main.go for now.
> 
> To break down the CL into smaller parts you could send a CL just with the dmstore package and/or one with the silveringestion package where the DMStore is stubbed out. These are packages that can largely be tested independently of each other (for silveringestion you would need to make DMStore an interface) and don't need the main function right away.

Regarding your suggestions about the architecture, that's definitely the approach I had in mind at first, but I was confused as to how to pass in the same DiffStore that's instantiated in main.go (so that the images can be downloaded locally and available to the frontend) given that newSilverProcessor has to implement the ingestion.Constructor signature. I chose to put the DiffStore in DMStore simply because I was able to make my own function signature for the DMStore constructor, and easily make sense of how I would be calling Init in main to pass in the DiffStore created there and create the singleton DMStore. Perhaps I'm incorrect in thinking that it necessarily has to be the same DiffStore instance -- maybe as long as the local base directories used to instantiate two DiffStores are the same, the images downloaded by one will be available to the other's ImageHandler? But definitely think it makes more sense to only have one DiffStore that's doing all the work related to diffing, and it's unclear to me how this would be passed in by the server when the ingestion processor is created. Could you clarify how this might work?

Patch-set: 1
Reviewer: Lawrence Choi <7450@5086850b-e599-37d2-8ec8-ff16aeccbf02>