silveringestion package
silveringestion enables the creation of a silverProcessor, which implements the
ingestion.Processor interface and facilitates the continuous parsing of metadata
from GS, from which information about screenshots taken during CT pixel diff
runs can be extracted. A silverProcessor contains a boltDB instance that holds
information about screenshots taken: each run has its own bucket, and within a
bucket, there are records that have screenshot filenames as keys and a list of
two bools as values. The bools correspond to whether the nopatch and withpatch
images for a particular screenshot filename have been processed. Once both are
marked true, the processor uses its DiffStore to calculate diff metrics for the
two images; in the process, the DiffStore also downloads the nopatch, withpatch,
and diff images to the local filesystem. These local image paths will be used by
the new server to render the images corresponding to a pixel diff run on the
frontend.

Bug: skia:6779
Change-Id: Ib5fa22a5f335147275c1fd0c6aa2fc244734fed4