dmstore package
dmstore enables the creation of a DMStore object, which is essentially just a
wrapper around a boltDB instance used to store diff metrics. The Add function
takes in a CT pixel diff run ID, URL screenshot filename, and diff.DiffMetrics
object: the  runID is used to create a bucket in the boltDB instance, and within
a bucket, the records have screenshot filenames as keys and serialized
DiffMetrics objects as values. The Remove function takes in a runID to remove
the bucket corresponding to that run from the boltDB instance. dmstore will be
used by silveringestion to store diff metrics after they are calculated.

Bug: skia:6779
Change-Id: I06237c7716244ca9a343025589540a75a99eaa97