Update patch set 1
Patch Set 1:

While I was writing this code, I was wondering if it'd be useful to create a go.skia.org/infra/go/time package with most of the utilities from the built-in time package, but with a testutils sub-package which allowed mocking the current time as done in the counters_test.  Maybe I'll write that code if I need this type of thing again.

Patch-set: 1
Reviewer: Ben Wagner <5120@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Reviewer: Eric Boren <5015@5086850b-e599-37d2-8ec8-ff16aeccbf02>