resultstore package
ResultStore is an interface for storing results extracted from Cluster Telemetry
Pixel Diff JSON metadata. Its records are ResultRec structs, which contain
information about a pair of nopatch and withpatch screenshots, and their diff
metrics. A ResultStore has three functions: Get, which uses a runID and url to
retrieve a ResultRec, Add, which places a ResultRec into the ResultRec using
the runID and url as keys, and RemoveRun, which removes all the ResultRecs
associated with the given runID.

BoltResultStore is an implementation of ResultStore that uses a boltDB instance
for storage. Within the boltDB instance, each runID has its own bucket, and
within a bucket, there are records that have screenshot urls as keys and
serialized ResultRecs as values.

Bug: skia:6779
Change-Id: I06237c7716244ca9a343025589540a75a99eaa97